Adding timeout impl that will do thread dump on expiration, and refine bmunit rules. Split up bmunit tests and run them with redirected output.
