+++ src/test/java/org/commonjava/util/partyline/fixture/ThreadDumper.java
@@ -1,18 +1,25 @@
 package org.commonjava.util.partyline.fixture;
 
-import org.apache.commons.lang.StringUtils;
 import org.junit.rules.TestName;
 import org.junit.rules.TestRule;
 import org.junit.rules.Timeout;
 import org.junit.runners.model.Statement;
+import org.junit.runners.model.TestTimedOutException;
 
 import java.lang.management.ManagementFactory;
 import java.lang.management.MonitorInfo;
 import java.lang.management.ThreadInfo;
 import java.lang.management.ThreadMXBean;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.FutureTask;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Stream;
 
 import static org.apache.commons.lang.StringUtils.join;
+import static org.junit.Assert.fail;
 
 /**
  * Created by jdcasey on 11/28/16.
@@ -44,47 +51,88 @@ public final class ThreadDumper
             MonitorInfo[] monitors = ti.getLockedMonitors();
             if ( monitors == null || monitors.length < 1 )
             {
-                sb.append("  -NONE-");
+                sb.append( "  -NONE-" );
             }
             else
             {
                 sb.append( "\n  - " ).append( join( monitors, "\n  - " ) );
             }
 
-            sb.append( "\n  Trace:\n    " )
-              .append( join( ti.getStackTrace(), "\n    " ) );
+            sb.append( "\n  Trace:\n    " ).append( join( ti.getStackTrace(), "\n    " ) );
 
         } );
 
         System.out.println( sb );
     }
 
-    public static TestRule wrap( TestName name, Timeout build )
+    public static TestRule timeoutRule( int timeout, TimeUnit units )
     {
-        return ( base, description ) -> new ThreadDumpWrapper( name, build.apply( base, description ) );
-    }
-
-    private static class ThreadDumpWrapper
-            extends Statement
-    {
-        private final String testMethod;
-
-        private Statement base;
-
-        public ThreadDumpWrapper( TestName name, Statement base )
+        return ( base, description ) -> new Statement()
         {
-            this.testMethod = name.getMethodName();
-            this.base = base;
-        }
+            public void evaluate()
+                    throws Throwable
+            {
+                System.out.printf( "Setting up timeout: %d %s to wrap: %s\n", timeout, units, base );
+                AtomicReference<Throwable> error = new AtomicReference<>();
+                CountDownLatch latch = new CountDownLatch( 1 );
+                FutureTask<Void> task = new FutureTask<>( () -> {
+                    try
+                    {
+                        latch.countDown();
+                        base.evaluate();
+                    }
+                    catch ( Throwable t )
+                    {
+                        error.set( t );
+                    }
 
-        @Override
-        public void evaluate()
-                throws Throwable
-        {
-            System.out.println( "Test timeout: " + testMethod + ". Thread dump: ");
-            dumpThreads();
+                    return null;
+                } );
 
-            base.evaluate();
-        }
+                ThreadGroup tg = new ThreadGroup( "Test Timeout Group" );
+                Thread t = new Thread( tg, task, "Test Timeout Thread" );
+                t.setDaemon( true );
+                t.start();
+
+                try
+                {
+                    System.out.println("Waiting for test to start.");
+                    latch.await();
+                }
+                catch ( InterruptedException e )
+                {
+                    error.set( e );
+                }
+
+                if ( error.get() == null )
+                {
+                    try
+                    {
+                        System.out.println( "Waiting for test to complete (or timeout)" );
+                        task.get( timeout, units );
+                    }
+                    catch ( InterruptedException e )
+                    {
+                        error.set( e );
+                    }
+                    catch ( ExecutionException e )
+                    {
+                        error.set( e.getCause() );
+                    }
+                    catch ( TimeoutException e )
+                    {
+                        System.out.printf( "Test timeout %d %s expired!\n", timeout, units.name() );
+                        dumpThreads();
+                        fail( String.format( "Test timed out after %d %s", timeout, units ) );
+                    }
+                }
+
+                Throwable throwable = error.get();
+                if ( throwable != null )
+                {
+                    throw throwable;
+                }
+            }
+        };
     }
 }
